Moved interfaces/abstract classes to another project
this was done to avoid installing statsd client dependency in all the
projects.
