+++ Ve.Metrics.StatsDClient.Abstract/IStatsdConfig.cs
@@ -1,0 +1,10 @@
+﻿namespace Ve.Metrics.StatsDClient.Abstract
+{
+    public interface IStatsdConfig
+    {
+        string Host { get; }
+        string Datacenter { get; }
+        int Port { get; }
+        string AppName { get; }
+    }
+}
