+++ Ve.Metrics.StatsDClient.Abstract/ITimingToken.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Ve.Metrics.StatsDClient.Abstract
+{
+    public interface ITimingToken : IDisposable {  }
+}
