+++ Ve.Metrics.StatsDClient.Abstract/SystemMetrics/SystemMetric.cs
@@ -1,0 +1,8 @@
+﻿namespace Ve.Metrics.StatsDClient.Abstract.SystemMetrics
+{
+    public abstract class SystemMetric
+    {
+        public abstract string Name { get; }
+        public abstract void Execute(IVeStatsDClient client);
+    }
+}
