+++ Ve.Metrics.StatsDClient.CastleWindsor/BaseInterceptor.cs
@@ -1,5 +1,7 @@
 ﻿using System.Linq;
 using Castle.DynamicProxy;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
 using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.CastleWindsor
