+++ Ve.Metrics.StatsDClient.CastleWindsor/StatsDCountingInterceptor.cs
@@ -1,4 +1,5 @@
 ﻿using Castle.DynamicProxy;
+using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.CastleWindsor
