+++ Ve.Metrics.StatsDClient.CastleWindsor/StatsDTimingInterceptor.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using Castle.DynamicProxy;
+using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.CastleWindsor
