+++ Ve.Metrics.StatsDClient.SimpleInjector/StatsDCountingInterceptor.cs
@@ -1,4 +1,5 @@
-﻿using Ve.Metrics.StatsDClient.Attributes;
+﻿using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.SimpleInjector
 {
