+++ Ve.Metrics.StatsDClient.SimpleInjector/StatsDTimingInterceptor.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.SimpleInjector
