+++ Ve.Metrics.StatsDClient.Tests/CastleWindsor/WindsorInterceptorsTestBase.cs
@@ -3,6 +3,8 @@ using Castle.MicroKernel.Registration;
 using Castle.Windsor;
 using Moq;
 using NUnit.Framework;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
 using Ve.Metrics.StatsDClient.Attributes;
 using Ve.Metrics.StatsDClient.CastleWindsor;
