+++ Ve.Metrics.StatsDClient.Tests/SimpleInjector/SimpleInjectorInterceptorsTestBase.cs
@@ -1,6 +1,8 @@
 using Moq;
 using NUnit.Framework;
 using SimpleInjector;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
 using Ve.Metrics.StatsDClient.Attributes;
 using Ve.Metrics.StatsDClient.SimpleInjector;
