+++ Ve.Metrics.StatsDClient.Tests/SystemMetrics/SystemMetricsExecutorShould.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 using System.Timers;
 using Moq;
 using NUnit.Framework;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
+using Ve.Metrics.StatsDClient.Abstract.SystemMetrics;
 using Ve.Metrics.StatsDClient.SystemMetrics;
 
 namespace Ve.Metrics.StatsDClient.Tests.SystemMetrics
