+++ Ve.Metrics.StatsDClient.Tests/Unity/UnityInterceptorsTestBase.cs
@@ -2,6 +2,8 @@ using Microsoft.Practices.Unity;
 using Microsoft.Practices.Unity.InterceptionExtension;
 using Moq;
 using NUnit.Framework;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
 using Ve.Metrics.StatsDClient.Attributes;
 using Ve.Metrics.StatsDClient.Unity;
