+++ Ve.Metrics.StatsDClient.Unity/StatsDCountingInterceptor.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.Practices.Unity.InterceptionExtension;
+using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Attributes;
 
 namespace Ve.Metrics.StatsDClient.Unity
