+++ Ve.Metrics.StatsDClient.WebApi/Ve.Metrics.StatsDClient.WebApi.csproj
@@ -64,6 +64,10 @@
     </None>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Ve.Metrics.StatsDClient.Abstract\Ve.Metrics.StatsDClient.Abstract.csproj">
+      <Project>{05e3c175-bdfa-4da6-bd01-4230fc668d74}</Project>
+      <Name>Ve.Metrics.StatsDClient.Abstract</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Ve.Metrics.StatsDClient\Ve.Metrics.StatsDClient.csproj">
       <Project>{3B4CCBBF-0F57-4D8F-9127-CF80203D0527}</Project>
       <Name>Ve.Metrics.StatsDClient</Name>
