+++ Ve.Metrics.StatsDClient.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ve.Metrics.StatsDClient.WebApi", "Ve.Metrics.StatsDClient.WebApi\Ve.Metrics.StatsDClient.WebApi.csproj", "{2C7F3091-E9B6-41FC-823B-347426BA0817}"
 EndProject
@@ -16,6 +16,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ve.Metrics.StatsDClient.Unity", "Ve.Metrics.StatsDClient.Unity\Ve.Metrics.StatsDClient.Unity.csproj", "{00CA99DB-703F-4C1A-A5D1-869A815766AC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ve.Metrics.StatsDClient.Mvc", "Ve.Metrics.StatsDClient.Mvc\Ve.Metrics.StatsDClient.Mvc.csproj", "{4010110C-043C-4DC9-894C-2D682E963038}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ve.Metrics.StatsDClient.Abstract", "Ve.Metrics.StatsDClient.Abstract\Ve.Metrics.StatsDClient.Abstract.csproj", "{05E3C175-BDFA-4DA6-BD01-4230FC668D74}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,6 +53,10 @@ Global
 		{4010110C-043C-4DC9-894C-2D682E963038}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4010110C-043C-4DC9-894C-2D682E963038}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4010110C-043C-4DC9-894C-2D682E963038}.Release|Any CPU.Build.0 = Release|Any CPU
+		{05E3C175-BDFA-4DA6-BD01-4230FC668D74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{05E3C175-BDFA-4DA6-BD01-4230FC668D74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{05E3C175-BDFA-4DA6-BD01-4230FC668D74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{05E3C175-BDFA-4DA6-BD01-4230FC668D74}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
