--- Ve.Metrics.StatsDClient/Attributes/StatsDBase.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Ve.Metrics.StatsDClient.Attributes
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    public abstract class StatsDBase : Attribute
-    {
-        public string Name { get; set; }
-        public Dictionary<string, string> Tags { get; set; }
-
-        protected StatsDBase(string name)
-        {
-            Name = name;
-        }
-    }
-}
