+++ Ve.Metrics.StatsDClient/Attributes/StatsDCounting.cs
@@ -1,4 +1,6 @@
-﻿namespace Ve.Metrics.StatsDClient.Attributes
+﻿using Ve.Metrics.StatsDClient.Abstract.Attributes;
+
+namespace Ve.Metrics.StatsDClient.Attributes
 {
     public class StatsDCounting : StatsDBase
     {
