+++ Ve.Metrics.StatsDClient/Attributes/StatsDTiming.cs
@@ -1,4 +1,6 @@
-﻿namespace Ve.Metrics.StatsDClient.Attributes
+﻿using Ve.Metrics.StatsDClient.Abstract.Attributes;
+
+namespace Ve.Metrics.StatsDClient.Attributes
 {
     public class StatsDTiming : StatsDBase
     {
