--- Ve.Metrics.StatsDClient/SystemMetrics/SystemMetric.cs
@@ -1,8 +1,0 @@
-﻿namespace Ve.Metrics.StatsDClient.SystemMetrics
-{
-    public abstract class SystemMetric
-    {
-        public abstract string Name { get; }
-        public abstract void Execute(IVeStatsDClient client);
-    }
-}
