+++ Ve.Metrics.StatsDClient/SystemMetrics/SystemMetricsExecutor.cs
@@ -1,5 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Timers;
+using Ve.Metrics.StatsDClient.Abstract;
+using Ve.Metrics.StatsDClient.Abstract.Attributes;
+using Ve.Metrics.StatsDClient.Abstract.SystemMetrics;
 
 namespace Ve.Metrics.StatsDClient.SystemMetrics
 {
