+++ Ve.Metrics.StatsDClient/VeStatsDClient.cs
@@ -2,23 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using StatsdClient;
+using Ve.Metrics.StatsDClient.Abstract;
 
 namespace Ve.Metrics.StatsDClient
 {
-    public interface IVeStatsDClient
-    {
-        void LogCount(string name);
-        void LogCount(string name, Dictionary<string, string> tags);
-        void LogCount(string name, int count, Dictionary<string, string> tags = null);
-        void LogTiming(string name, long milliseconds, Dictionary<string, string> tags = null);
-        void LogTiming(string name, int milliseconds, Dictionary<string, string> tags = null);
-        TimingToken LogTiming(string name);
-        void LogGauge(string name, int value, Dictionary<string, string> tags = null);
-        void LogCalendargram(string name, int value, string period, Dictionary<string, string> tags = null);
-        void LogCalendargram(string name, string value, string period, Dictionary<string, string> tags = null);
-        void LogRaw(string name, int value, string period, long? epoch = null, Dictionary<string, string> tags = null);
-    }
-
     public class VeStatsDClient : IVeStatsDClient
     {
         private static string _systemTags;
@@ -68,7 +55,7 @@ namespace Ve.Metrics.StatsDClient
             _statsd.LogTiming(BuildName(name, tags), milliseconds);
         }
 
-        public TimingToken LogTiming(string name)
+        public ITimingToken LogTiming(string name)
         {
             return new TimingToken(this, name);
         }
